+++ mihomo/listener/sing/util.go
@@ -12,11 +12,14 @@ import (
 func (h *ListenerHandler) HandleSocket(target socks5.Addr, conn net.Conn, _additions ...inbound.Addition) {
 	conn, metadata := inbound.NewSocket(target, conn, h.Type, h.Additions...)
 	if h.IsSpecialFqdn(metadata.Host) {
-		_ = h.ParseSpecialFqdn(
+		err := h.ParseSpecialFqdn(
 			WithAdditions(context.Background(), _additions...),
 			conn,
 			ConvertMetadata(metadata),
 		)
+		if err != nil {
+			_ = conn.Close()
+		}
 	} else {
 		inbound.ApplyAdditions(metadata, _additions...)
 		h.Tunnel.HandleTCPConn(conn, metadata)