+++ sing-box/common/ktls/ktls.go
@@ -3,6 +3,7 @@
 package ktls
 
 import (
+	"context"
 	"crypto/tls"
 	"io"
 	"net"
@@ -10,24 +11,30 @@ import (
 	"syscall"
 
 	"github.com/sagernet/sing-box/common/badtls"
-	// C "github.com/sagernet/sing-box/constant"
 	E "github.com/sagernet/sing/common/exceptions"
+	"github.com/sagernet/sing/common/logger"
 	N "github.com/sagernet/sing/common/network"
 	aTLS "github.com/sagernet/sing/common/tls"
 )
 
 type Conn struct {
 	aTLS.Conn
+	ctx             context.Context
+	logger          logger.ContextLogger
 	conn            net.Conn
 	rawConn         *badtls.RawConn
+	syscallConn     syscall.Conn
 	rawSyscallConn  syscall.RawConn
 	readWaitOptions N.ReadWaitOptions
 	kernelTx        bool
 	kernelRx        bool
-	tmp             [16]byte
 }
 
-func NewConn(conn aTLS.Conn, txOffload, rxOffload bool) (aTLS.Conn, error) {
+func NewConn(ctx context.Context, logger logger.ContextLogger, conn aTLS.Conn, txOffload, rxOffload bool) (aTLS.Conn, error) {
+	err := Load()
+	if err != nil {
+		return nil, err
+	}
 	syscallConn, isSyscallConn := N.CastReader[interface {
 		io.Reader
 		syscall.Conn
@@ -54,14 +61,17 @@ func NewConn(conn aTLS.Conn, txOffload, rxOffload bool) (aTLS.Conn, error) {
 		for rawConn.Hand.Len() > 0 {
 			err = rawConn.HandlePostHandshakeMessage()
 			if err != nil {
-				return nil, E.Cause(err, "ktls: failed to handle post-handshake messages")
+				return nil, E.Cause(err, "handle post-handshake messages")
 			}
 		}
 	}
 	kConn := &Conn{
 		Conn:           conn,
+		ctx:            ctx,
+		logger:         logger,
 		conn:           conn.NetConn(),
 		rawConn:        rawConn,
+		syscallConn:    syscallConn,
 		rawSyscallConn: rawSyscallConn,
 	}
 	err = kConn.setupKernel(txOffload, rxOffload)
@@ -72,13 +82,25 @@ func NewConn(conn aTLS.Conn, txOffload, rxOffload bool) (aTLS.Conn, error) {
 }
 
 func (c *Conn) Upstream() any {
-	return c.conn
+	return c.Conn
 }
 
-func (c *Conn) ReaderReplaceable() bool {
-	return c.kernelRx
+func (c *Conn) SyscallConnForRead() syscall.Conn {
+	if !c.kernelRx {
+		return nil
+	}
+	if !*c.rawConn.IsClient {
+		c.logger.WarnContext(c.ctx, "ktls: RX splice is unavailable on the server size, since it will cause an unknown failure")
+		return nil
+	}
+	c.logger.DebugContext(c.ctx, "ktls: RX splice requested")
+	return c.syscallConn
 }
 
-func (c *Conn) WriterReplaceable() bool {
-	return c.kernelTx
+func (c *Conn) SyscallConnForWrite() syscall.Conn {
+	if !c.kernelTx {
+		return nil
+	}
+	c.logger.DebugContext(c.ctx, "ktls: TX splice requested")
+	return c.syscallConn
 }