+++ sing-box/common/ktls/ktls_stub.go
@@ -3,11 +3,13 @@
 package ktls
 
 import (
+	"context"
 	"os"
 
+	"github.com/sagernet/sing/common/logger"
 	aTLS "github.com/sagernet/sing/common/tls"
 )
 
-func NewConn(conn aTLS.Conn, txOffload, rxOffload bool) (aTLS.Conn, error) {
+func NewConn(ctx context.Context, logger logger.ContextLogger, conn aTLS.Conn, txOffload, rxOffload bool) (aTLS.Conn, error) {
 	return nil, os.ErrInvalid
 }