+++ sing-box/common/tls/config.go
@@ -21,12 +21,6 @@ type (
 	CurveID         = tls.CurveID
 )
 
-type KTLSCapableConfig interface {
-	Config
-	KernelTx() bool
-	KernelRx() bool
-}
-
 func ParseTLSVersion(version string) (uint16, error) {
 	switch version {
 	case "1.0":